Move post-processing functions for features into new module
